+++ lnbits/core/services/__init__.py
@@ -12,7 +12,6 @@ from .payments import (
     fee_reserve,
     fee_reserve_total,
     pay_invoice,
-    send_payment_notification,
     service_fee,
     update_pending_payments,
     update_wallet_balance,
@@ -48,7 +47,6 @@ __all__ = [
     "fee_reserve",
     "fee_reserve_total",
     "pay_invoice",
-    "send_payment_notification",
     "service_fee",
     "update_pending_payments",
     "update_wallet_balance",
