+++ lnbits/core/services/payments.py
@@ -1,5 +1,4 @@
 import asyncio
-import json
 import time
 from datetime import datetime, timedelta, timezone
 from typing import Optional
@@ -12,9 +11,7 @@ from loguru import logger
 from lnbits.core.crud.payments import get_daily_stats
 from lnbits.core.db import db
 from lnbits.core.models import PaymentDailyStats, PaymentFilters
-from lnbits.core.models.notifications import NotificationType
-from lnbits.core.services.notifications import enqueue_notification
-from lnbits.core.tasks import dispatch_webhook, send_payment_push_notification
+from lnbits.core.tasks import send_payment_notification
 from lnbits.db import Connection, Filters
 from lnbits.decorators import check_user_extension_access
 from lnbits.exceptions import InvoiceError, PaymentError
@@ -46,7 +43,6 @@ from ..models import (
     PaymentState,
     Wallet,
 )
-from .websockets import websocket_manager
 
 
 async def pay_invoice(
@@ -277,67 +273,6 @@ async def update_wallet_balance(
         from lnbits.tasks import internal_invoice_queue
 
         await internal_invoice_queue.put(payment.checking_id)
-
-
-async def send_payment_notification(wallet: Wallet, payment: Payment):
-    try:
-        await send_ws_payment_notification(wallet, payment)
-    except Exception as e:
-        logger.error("Error sending websocket payment notification", e)
-
-    try:
-        send_chat_payment_notification(wallet, payment)
-    except Exception as e:
-        logger.error("Error sending chat payment notification", e)
-    try:
-        await send_payment_push_notification(payment)
-    except Exception as e:
-        logger.error("Error sending push payment notification", e)
-
-    if payment.webhook and not payment.webhook_status:
-        await dispatch_webhook(payment)
-
-
-async def send_ws_payment_notification(wallet: Wallet, payment: Payment):
-    # TODO: websocket message should be a clean payment model
-    # await websocket_manager.send_data(payment.json(), wallet.inkey)
-    # TODO: figure out why we send the balance with the payment here.
-    # cleaner would be to have a separate message for the balance
-    # and send it with the id of the wallet so wallets can subscribe to it
-    payment_notification = json.dumps(
-        {
-            "wallet_balance": wallet.balance,
-            # use pydantic json serialization to get the correct datetime format
-            "payment": json.loads(payment.json()),
-        },
-    )
-    await websocket_manager.send_data(payment_notification, wallet.inkey)
-    await websocket_manager.send_data(payment_notification, wallet.adminkey)
-
-    await websocket_manager.send_data(
-        json.dumps({"pending": payment.pending}), payment.payment_hash
-    )
-
-
-def send_chat_payment_notification(wallet: Wallet, payment: Payment):
-    amount_sats = abs(payment.sat)
-    values: dict = {
-        "wallet_id": wallet.id,
-        "wallet_name": wallet.name,
-        "amount_sats": amount_sats,
-        "fiat_value_fmt": "",
-    }
-    if payment.extra.get("wallet_fiat_currency", None):
-        amount_fiat = payment.extra.get("wallet_fiat_amount", None)
-        currency = payment.extra.get("wallet_fiat_currency", None)
-        values["fiat_value_fmt"] = f"`{amount_fiat}`*{currency}* / "
-
-    if payment.is_out:
-        if amount_sats >= settings.lnbits_notification_outgoing_payment_amount_sats:
-            enqueue_notification(NotificationType.outgoing_payment, values)
-    else:
-        if amount_sats >= settings.lnbits_notification_incoming_payment_amount_sats:
-            enqueue_notification(NotificationType.incoming_payment, values)
 
 
 async def check_wallet_limits(
