+++ lnbits/core/tasks.py
@@ -1,4 +1,5 @@
 import asyncio
+import json
 import traceback
 from typing import Callable, Coroutine
 
@@ -15,12 +16,9 @@ from lnbits.core.crud.audit import delete_expired_audit_entries
 from lnbits.core.crud.payments import get_payments_status_count
 from lnbits.core.crud.users import get_accounts
 from lnbits.core.crud.wallets import get_wallets_count
-from lnbits.core.models import AuditEntry, Payment
+from lnbits.core.models import AuditEntry, Payment, Wallet
 from lnbits.core.models.extensions import InstallableExtension
 from lnbits.core.models.notifications import NotificationType
-from lnbits.core.services import (
-    send_payment_notification,
-)
 from lnbits.core.services.funding_source import (
     check_balance_delta_changed,
     check_server_balance_against_node,
@@ -30,6 +28,7 @@ from lnbits.core.services.notifications import (
     enqueue_notification,
     process_next_notification,
 )
+from lnbits.core.services.websockets import websocket_manager
 from lnbits.db import Filters
 from lnbits.helpers import check_callback_url
 from lnbits.settings import settings
@@ -134,6 +133,66 @@ async def dispatch_webhook(payment: Payment):
         except httpx.RequestError:
             await mark_webhook_sent(payment.payment_hash, -1)
             logger.warning(f"Could not send webhook to {payment.webhook}")
+
+
+async def send_payment_notification(wallet: Wallet, payment: Payment):
+    try:
+        await send_ws_payment_notification(wallet, payment)
+    except Exception as e:
+        logger.error("Error sending websocket payment notification", e)
+    try:
+        send_chat_payment_notification(wallet, payment)
+    except Exception as e:
+        logger.error("Error sending chat payment notification", e)
+    try:
+        await send_payment_push_notification(payment)
+    except Exception as e:
+        logger.error("Error sending push payment notification", e)
+
+    if payment.webhook and not payment.webhook_status:
+        await dispatch_webhook(payment)
+
+
+async def send_ws_payment_notification(wallet: Wallet, payment: Payment):
+    # TODO: websocket message should be a clean payment model
+    # await websocket_manager.send_data(payment.json(), wallet.inkey)
+    # TODO: figure out why we send the balance with the payment here.
+    # cleaner would be to have a separate message for the balance
+    # and send it with the id of the wallet so wallets can subscribe to it
+    payment_notification = json.dumps(
+        {
+            "wallet_balance": wallet.balance,
+            # use pydantic json serialization to get the correct datetime format
+            "payment": json.loads(payment.json()),
+        },
+    )
+    await websocket_manager.send_data(payment_notification, wallet.inkey)
+    await websocket_manager.send_data(payment_notification, wallet.adminkey)
+
+    await websocket_manager.send_data(
+        json.dumps({"pending": payment.pending}), payment.payment_hash
+    )
+
+
+def send_chat_payment_notification(wallet: Wallet, payment: Payment):
+    amount_sats = abs(payment.sat)
+    values: dict = {
+        "wallet_id": wallet.id,
+        "wallet_name": wallet.name,
+        "amount_sats": amount_sats,
+        "fiat_value_fmt": "",
+    }
+    if payment.extra.get("wallet_fiat_currency", None):
+        amount_fiat = payment.extra.get("wallet_fiat_amount", None)
+        currency = payment.extra.get("wallet_fiat_currency", None)
+        values["fiat_value_fmt"] = f"`{amount_fiat}`*{currency}* / "
+
+    if payment.is_out:
+        if amount_sats >= settings.lnbits_notification_outgoing_payment_amount_sats:
+            enqueue_notification(NotificationType.outgoing_payment, values)
+    else:
+        if amount_sats >= settings.lnbits_notification_incoming_payment_amount_sats:
+            enqueue_notification(NotificationType.incoming_payment, values)
 
 
 async def send_payment_push_notification(payment: Payment):
